Add files via upload
Take home assignments from the course. 

Initially we were told that we would be graded on 2 assignments out of 3. It so happened that only 2 assignments were given through the semester, so as an extension of the project in which we were to build a model of voyager's slingshot around jupiter, I had to build a 3 body gravitational model as an extra credit, second assignment.